scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Scripts for doodlesdev.com"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["pyyaml>=6.0.3", "requests>=2.32.5"]
